# docs/swagger.yaml
basePath: /
definitions:
  internal_delivery_http.FunnelRequest:
    properties:
      background:
        type: boolean
      port:
        type: integer
    required:
    - port
    type: object
  internal_delivery_http.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  internal_delivery_http.ServeRequest:
    properties:
      background:
        type: boolean
      port:
        type: integer
    required:
    - port
    type: object
host: localhost:8080
info:
  contact: {}
  description: Self-contained HTTP API for Tailscale with embedded daemon
  title: Tailscale Controller API
  version: "2.0"
paths:
  /api/v1/funnel:
    delete:
      description: Reset tailscale funnel configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Reset funnel
      tags:
      - tailscale
    post:
      consumes:
      - application/json
      description: Start tailscale funnel on specified port
      parameters:
      - description: Funnel configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.FunnelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Start funnel
      tags:
      - tailscale
  /api/v1/funnel/status:
    get:
      description: Get current tailscale funnel configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Get funnel status
      tags:
      - tailscale
  /api/v1/logs/app:
    get:
      description: Get application log file content
      parameters:
      - default: 100
        description: Number of lines to retrieve
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Get application logs
      tags:
      - logs
  /api/v1/serve:
    delete:
      description: Reset tailscale serve configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Reset serve
      tags:
      - tailscale
    post:
      consumes:
      - application/json
      description: Start tailscale serve on specified port
      parameters:
      - description: Serve configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.ServeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Start serve
      tags:
      - tailscale
  /api/v1/serve/status:
    get:
      description: Get current tailscale serve configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Get serve status
      tags:
      - tailscale
  /api/v1/ssh/enable:
    post:
      description: Enable tailscale SSH access
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Enable SSH
      tags:
      - tailscale
  /api/v1/status:
    get:
      description: Get current tailscale connection status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.Response'
      summary: Get tailscale status
      tags:
      - tailscale
swagger: "2.0"